Revert "server broken"
This reverts commit 0416fe807c8dd17d7d1276c9bf5dec21f40d4169.
