--- Code/infochange/server/controller/user.js
@@ -1,3 +1,0 @@
-const controller = {};
-
-module.exports = controller;
