+++ Code/infochange/server/package.json
@@ -3,8 +3,8 @@
     "version": "1.0.0",
     "main": "src/index.js",
     "scripts": {
-        "start": "node index.js",
-        "dev": "node --watch index.js",
+        "start": "node src/index.js",
+        "dev": "node --watch src/index.js",
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "keywords": [],
